알림(차단) 도메인 리펙토링
