+++ src/main/java/com/shelfeed/backend/domain/notification/service/NotificationService.java
@@ -1,7 +1,5 @@
 package com.shelfeed.backend.domain.notification.service;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.shelfeed.backend.domain.member.entity.Member;
 import com.shelfeed.backend.domain.member.repository.MemberRepository;
 import com.shelfeed.backend.domain.notification.dto.response.NotificationItemResponse;
@@ -11,13 +9,12 @@ import com.shelfeed.backend.domain.notification.entity.Notification;
 import com.shelfeed.backend.domain.notification.repository.NotificationRepository;
 import com.shelfeed.backend.global.common.exception.BusinessException;
 import com.shelfeed.backend.global.common.exception.ErrorCode;
+import com.shelfeed.backend.global.common.util.CursorUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.nio.charset.StandardCharsets;
-import java.util.Base64;
 import java.util.List;
 
 @Service
@@ -27,7 +24,7 @@ public class NotificationService {
 
     private final NotificationRepository notificationRepository;
     private final MemberRepository memberRepository;
-    private final ObjectMapper objectMapper = new ObjectMapper();
+    private final CursorUtils cursorUtils;
 
     public NotificationListResponse getMyNotifications(Long memberUserId, String cursor, int limit) {
          if (limit <= 0) {
@@ -36,7 +33,7 @@ public class NotificationService {
         Member receiver = memberRepository.findByMemberUserId(memberUserId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.MEMBER_NOT_FOUND));
 
-        Long cursorId = decodeCursor(cursor);
+        Long cursorId = cursorUtils.decode(cursor);
         List<Notification> notifications = notificationRepository.findMyNotifications(
                 receiver,
                 cursorId,
@@ -50,7 +47,7 @@ public class NotificationService {
         String nextCursor = null;
         if (content.size() > limit) {
             Long nextId = content.get(limit - 1).getNotificationId();
-            nextCursor = encodeCursor(nextId);
+            nextCursor = cursorUtils.encode(nextId);
         }
 
         return NotificationListResponse.of(content, limit, nextCursor);
@@ -76,28 +73,5 @@ public class NotificationService {
         return UnreadCountResponse.of(unreadCount);
     }
 
-    private Long decodeCursor(String cursor) {
-        if (cursor == null || cursor.isBlank()) return null;
-
-        try {
-            byte[] decoded = Base64.getDecoder().decode(cursor);
-            String json = new String(decoded, StandardCharsets.UTF_8);
-            CursorPayload payload = objectMapper.readValue(json, CursorPayload.class);
-            return payload.id();
-        } catch (IllegalArgumentException | JsonProcessingException e) {
-            throw new BusinessException(ErrorCode.INVALID_INPUT);
-        }
-    }
-
-    private String encodeCursor(Long id) {
-        try {
-            String json = objectMapper.writeValueAsString(new CursorPayload(id));
-            return Base64.getEncoder().encodeToString(json.getBytes(StandardCharsets.UTF_8));
-        } catch (JsonProcessingException e) {
-            throw new BusinessException(ErrorCode.INVALID_INPUT);
-        }
-    }
-
-    private record CursorPayload(Long id) {}
 }
